feat: Add openBrowser function to AuthServerProvider
The `AuthServerProvider` class now includes a new static method `setOpenBrowser` which allows the developer to set a custom function for opening the browser. This function will be used in the `openAuthDapp` method to open the browser on the auth dapp requests page. This change improves flexibility and allows for easier customization of the browser opening behavior.